# kratos-learn/urlshorter/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: User API
    version: 0.0.1
paths:
    /admin/user/login:
        post:
            tags:
                - User
            operationId: User_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.user.v1.UserRegisterLoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.user.v1.UserRegisterLoginResponse'
    /admin/user/register:
        post:
            tags:
                - User
            operationId: User_Register
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.user.v1.UserRegisterLoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.user.v1.UserRegisterLoginResponse'
components:
    schemas:
        api.user.v1.UserRegisterLoginRequest:
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
        api.user.v1.UserRegisterLoginResponse:
            type: object
            properties:
                statusCode:
                    type: integer
                    format: int32
                statusMsg:
                    type: string
                userId:
                    type: string
                token:
                    type: string
tags:
    - name: User